refactor(accounts): split 641-line views.py into focused package
Workstream C continuation. views.py handled four distinct concerns: auth
(login/register/logout/refresh), profile, staff (list/detail/activity),
and APP-12 data export. Split into accounts/views/ package with one file
per concern:

- auth.py           — CookieTokenRefreshView, RegisterView, LoginView, LogoutView
- profile.py        — UserProfileView, CustomerProfileView
- staff.py          — StaffCustomerListView, StaffCustomerProfileView, StaffCustomerActivityView
- export.py         — CustomerDataExportView
- _shared.py        — JWT cookie helpers + four rate-throttle classes
- __init__.py       — re-exports all public names (urls.py and tests.test_api_contracts reference via `from . import views` / `from apps.accounts.views import LoginView, RegisterView`, both unchanged)

Tests: 16 auth tests pass, 81 broader accounts-domain tests pass.
