+++ backend/apps/accounts/views/profile.py
@@ -1,0 +1,24 @@
+"""Profile views: current user and their customer profile."""
+
+from rest_framework import generics
+from rest_framework.permissions import IsAuthenticated
+
+from ..models import CustomerProfile
+from ..serializers import CustomerProfileSerializer, UserSerializer
+
+
+class UserProfileView(generics.RetrieveUpdateAPIView):
+    serializer_class = UserSerializer
+    permission_classes = (IsAuthenticated,)
+
+    def get_object(self):
+        return self.request.user
+
+
+class CustomerProfileView(generics.RetrieveUpdateAPIView):
+    serializer_class = CustomerProfileSerializer
+    permission_classes = (IsAuthenticated,)
+
+    def get_object(self):
+        profile, _ = CustomerProfile.objects.get_or_create(user=self.request.user)
+        return profile
